Manually Rebased to current COP-1 state.
