+++ oresat_c3/protocols/uslp.py
@@ -1,6 +1,12 @@
 from dataclasses import dataclass
+from typing import Optional
 
-from spacepackets.uslp import BypassSequenceControlFlag, PrimaryHeader, ProtocolCommandFlag
+from spacepackets.uslp import (
+    BypassSequenceControlFlag,
+    PrimaryHeader,
+    ProtocolCommandFlag,
+    SourceOrDestField,
+)
 from spacepackets.uslp.defs import UslpInvalidRawPacketOrFrameLenError
 from spacepackets.uslp.frame import (
     FrameType,
@@ -14,18 +20,17 @@ from spacepackets.uslp.frame import (
 SPACECRAFT_ID = 0x4F53  # aka "OS" in ASCII
 
 PRIMARY_HEADER_LEN = 7
-SPI_LEN = 2
 SEQ_NUM_LEN = 4
 DFH_LEN = 1
 HMAC_LEN = 32
 FECF_LEN = 2
-TC_MIN_LEN = PRIMARY_HEADER_LEN + SPI_LEN + SEQ_NUM_LEN + DFH_LEN + HMAC_LEN + FECF_LEN
+TC_MIN_LEN = PRIMARY_HEADER_LEN + DFH_LEN + FECF_LEN
 
 FRAME_PROPS = VarFrameProperties(
     has_insert_zone=True,
     has_fecf=True,
     truncated_frame_len=0,
-    insert_zone_len=0,#SPI_LEN + SEQ_NUM_LEN,
+    insert_zone_len=SEQ_NUM_LEN,
 )
 
 
@@ -116,17 +121,47 @@ def unpack_frame(raw: bytes) -> TransferFrame:
     return frame
 
 
-def pack(payload: bytes, seq_num: int, control_word: bytes) -> bytes:
-    tfdz = payload
+def make_frame(
+    payload: bytes,
+    vcid: int,
+    src_dest: SourceOrDestField,
+    vcf_count: Optional[int] = None,
+    control_word: Optional[bytes] = None,
+    insert_zone: Optional[bytes] = None,
+) -> TransferFrame:
+    """Create and pack a USLP
+
+    Parameters
+    ----------
+    payload
+        The pre-packed payload.
+    vcid
+        The Virtual Channel Identifier of the frame.
+    src_dest
+        The Source or Destination identifier.
+    vcf_count
+        The Virtual Channel Frame count. If None, the VCF length is to 0 and no count is specified.
+    control_word
+        The CLCW, if any, to pack in the frame.
+    insert_zone
+        The insert zone data, if any.
+
+    Returns
+    -------
+    TransferFrame
+        The constructed Transfer Frame.
+    """
 
     tfdf = TransferFrameDataField(
         tfdz_cnstr_rules=TfdzConstructionRules.VpNoSegmentation,
-        uslp_ident=UslpProtocolIdentifier.SPACE_PACKETS_ENCAPSULATION_PACKETS,
-        tfdz=tfdz,
+        uslp_ident=UslpProtocolIdentifier.USER_DEFINED_OCTET_STREAM,
+        tfdz=payload,
     )
 
     # USLP transfer frame total length - 1
-    frame_len = len(payload) + TC_MIN_LEN - 1 - HMAC_LEN - 6
+    frame_len = len(payload) + PRIMARY_HEADER_LEN + DFH_LEN + FECF_LEN - 1
+    if insert_zone:
+        frame_len += len(insert_zone)
 
     has_clcw = bool(control_word)
     if has_clcw:
@@ -135,20 +170,16 @@ def pack(payload: bytes, seq_num: int, control_word: bytes) -> bytes:
     frame_header = PrimaryHeader(
         scid=SPACECRAFT_ID,
         map_id=0,
-        vcid=0,
-        src_dest=0,
+        vcid=vcid,
+        src_dest=src_dest,
         frame_len=frame_len,
-        vcf_count_len=0,
+        vcf_count_len=bool(vcf_count),
+        vcf_count=vcf_count,
         op_ctrl_flag=has_clcw,
         prot_ctrl_cmd_flag=ProtocolCommandFlag.USER_DATA,
         bypass_seq_ctrl_flag=BypassSequenceControlFlag.SEQ_CTRLD_QOS,
     )
 
-    # sdls_header_bytes = int(1).to_bytes(2, "big")
-    # sdls_header_bytes += seq_num.to_bytes(SEQ_NUM_LEN, "big")
-    frame = TransferFrame(
-        header=frame_header, tfdf=tfdf, op_ctrl_field=control_word
+    return TransferFrame(
+        header=frame_header, tfdf=tfdf, op_ctrl_field=control_word, insert_zone=insert_zone
     )
-    packet = frame.pack(frame_type=FrameType.VARIABLE)
-
-    return packet
